// src/lib/api/openapi/models/LiliInfo.ts
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

export type LiliInfo = {
    id: string;
    name: string;
    n_links: number;
    description: (string | null);
    public_access: boolean;
    client_can_edit: boolean;
    edited_at: string;
    owner: string;
    profile_picture_url: (string | null);
};